create three folder to place static files and seperate routes from app.py using Flask Blueprints
