+++ static/css/booking.css
@@ -1,0 +1,164 @@
+body{
+    margin: 0px;
+    padding: 0px;
+}
+.grayscale{
+    display: none;
+    position: absolute;
+    top: 0%;
+    width: 100%;
+    height: 100px;
+    z-index: 9999;
+    background-color: #757575;
+    opacity: 0.4;
+}
+a{
+    text-decoration: none;
+}
+nav{  
+    align-items: center;
+    margin: 0 auto;
+    width: 1200px;
+    height: 54px;
+    display: flex;
+    justify-content: space-between;
+}
+.navbar_text{
+    color: #448899;
+    font-style: normal;
+    font-weight: 700;
+    font-size: 30px;
+}
+.reservation{
+    font-size: 16px;
+}
+.signin{
+    font-size: 16px;
+}
+.navbar_list > button{
+    background-color: #fff;
+    color: #000;
+    border: none;
+    cursor: pointer;
+    padding: 0px;
+}
+.navbar_list > button:hover{
+    transform: scale(1.06);
+}
+.navbar_list > button:active{
+    transform: scale(1);
+}
+main{
+    width: 1200px;
+    margin: 50px auto;
+
+}
+main >hr{
+    margin-top: 45px;
+    border: 0.5px solid #e8e8e8;
+}
+.attraction_img{
+    display: inline-block;
+    vertical-align: top;
+    margin-top: 35px;
+    margin-left: 110px;
+    width: 250px;
+    height: 200px;
+}
+.attraction_info{
+    display: inline-block;
+    margin-top: 35px;
+}
+.attraction_content{
+    display: inline-block;
+    margin: 15px 5px 0px 22px;
+    font-size: 16px;
+    font-family: 'Noto Sans TC';
+    font-weight: 700;
+    color: #666;
+}
+.attraction_title{
+    display: inline-block;
+    margin-left: 22px;
+    font-size: 16px;
+    font-family: 'Noto Sans TC';
+    font-weight: 700;
+    color: #448899;
+}
+span{
+    font-size: 16px;
+    font-family: 'Noto Sans TC';
+    font-weight: 500;
+    color: #666;
+}
+.title{
+    margin-top: 44px;
+    margin-left: 110px;
+    font-size: 19px;
+    font-family: 'Noto Sans TC';
+    font-weight: 700;
+    color: #666;
+}
+.content{
+    display: inline-block;
+    margin-top: 30px;
+    margin-left: 110px;
+    font-size: 16px;
+    font-family: 'Noto Sans TC';
+    font-weight: 500;
+    color: #666;
+}
+input{
+    box-sizing: border-box;
+    width: 200px;
+    height: 38px;
+    padding: 10px;
+    border: 1px solid #e8e8e8;
+    border-radius: 5px;
+    font-size: 16px;
+    color: #757575;
+}
+.money{
+    margin: 43px 110px 25px 0px;
+    font-size: 16px;
+    font-family: 'Noto Sans TC';
+    font-weight: 500;
+    color: #666;
+
+}
+main button{
+    width: 173px;
+    height: 36px;
+    margin-right: 110px;
+    padding: 10px 20px;
+    line-height: 19px;
+    background-color: #448899;
+    border: none;
+    border-radius: 5px;
+    font-size: 19px;
+    color: #fff;
+    cursor: pointer;
+}
+main button:hover{
+    transform: translateY(-2px);
+    box-shadow: 2px 2px 5px #000;
+}
+main button:active{
+    transform: translateY(0px);
+    box-shadow: none;
+}
+
+footer{
+    width: 100%;
+    height:104px;
+    background-color: #757575;
+    position: absolute;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+.copyright{
+    font-weight: 700;
+    font-size: 16px;
+    color: #fff;
+}
